fastapi/requirements.txt
alembic
fastapi[all]
fastapi-users[sqlalchemy]
asyncpg
python-dotenv
sqlalchemy
pydantic
psycopg2
PyJWT
boto3
botocore
loguru
python-magic-bin
